requirements.txt
datasets==1.1.2
hydra-core==1.0.3
ipython==7.16.1
matplotlib==3.3.3
numpy==1.19.3
omegaconf==2.0.4
plotly==4.14.3
pytorch-lightning==1.1.4
seaborn==0.11.1
tensorboard==2.3.0
tokenizers==0.9.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.7.1
tqdm==4.49.0
transformers==3.4.0
typing-extensions==3.7.4.3
wandb==0.10.17
